# improve Equals func performance 