@istio-testing 2021-06-21
@silenceshell: The following test **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
release-notes_istio | 37d4428c555a3a00cedf9cccb551f64e206ec7e0 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/33556/release-notes_istio/1407015430995841024) | `/test release-notes_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->