@silenceshell 2021-06-21
> Thanks for the PR! One note - https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go golang has pretty good native support for benchmarks that may help in the future to test your changes

Thanks for your advice!